Some Clean Up
Okay, so as far as I can tell, I'm working on the Platformer, Michael is
working on the Dungeon Crawler, and Zach is working on Space Invaders.
Do not work on any part of the game you're not assigned to. Caleb, Ian,
and Brayden, if you see this, could you work on the binder if you
haven't already?